+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+## Unreleased
+
+### Bug fixes
+
+* Only set the context to the current request URL if the context has not already been set
+  | [tremlab](https://github.com/tremlab)
+  | [#89](https://github.com/bugsnag/bugsnag-dotnet/pull/89)
+
+## 2.0.2 (2018-03-28)
+
+### Bug fixes
+
+* Allow the severity to be changed in callbacks
+  | [martin308](https://github.com/martin308)
+  | [#84](https://github.com/bugsnag/bugsnag-dotnet/pull/84)
+
 ## 2.0.1 (2018-03-27)
 
 ### Bug fixes